requirements.txt
pandas>=2.2.3
Pillow>=11.1.0
